transformations/requirements.txt
pytest~=6.2.4
boto3~=1.18.46
pandas~=1.3.3
python-dotenv~=0.19.2